The standard allows for a type that is effectively 'an int big enough to hold a pointer'.
Use that instead of 32 bit integer.  Fixes compile issues on Linux.

+++ hal/src/main/native/include/HAL/Types.h
@@ -11,7 +11,7 @@
 
 #define HAL_kInvalidHandle 0
 
-typedef int32_t HAL_Handle;
+typedef intptr_t HAL_Handle;
 
 typedef HAL_Handle HAL_PortHandle;
 
@@ -43,4 +43,4 @@ typedef HAL_Handle HAL_RelayHandle;
 
 typedef HAL_Handle HAL_SolenoidHandle;
 
-typedef int32_t HAL_Bool;
+typedef intptr_t HAL_Bool;
